# vtprotobuf: Reuse backing array with referenced structs instead of pooling individual objects